Add/refactor some unit tests
