+++ spec/lib/monitoring/query_helper_spec.rb
@@ -1,37 +1,46 @@
 require 'monitoring/query_helper'
 require 'spec_helper'
 
-describe Monitoring::QueryHelper, type: :request do
-  let(:queryhelper) { Monitoring::QueryHelper.instance }
+RSpec.describe Monitoring::QueryHelper do
+  describe '#policy_resource_counts' do
+    it 'returns the correct resource counts' do
+      allow(Resource).to receive(:group_and_count).and_return([
+        { kind: 'resource1', count: 10 },
+        { kind: 'resource2', count: 20 }
+      ])
 
-  let(:policies_url) { '/policies/rspec/policy/root' }
+      counts = Monitoring::QueryHelper.instance.policy_resource_counts
 
-  let(:current_user) { Role.find_or_create(role_id: 'rspec:user:admin') }
+      expect(counts).to eq({ 'resource1' => 10, 'resource2' => 20 })
+    end
 
-  let(:token_auth_header) do
-    bearer_token = Slosilo["authn:rspec"].signed_token(current_user.login)
-    token_auth_str =
-      "Token token=\"#{Base64.strict_encode64(bearer_token.to_json)}\""
-    { 'HTTP_AUTHORIZATION' => token_auth_str }
+    it 'returns an empty hash if there are no resources' do
+      allow(Resource).to receive(:group_and_count).and_return([])
+
+      counts = Monitoring::QueryHelper.instance.policy_resource_counts
+
+      expect(counts).to eq({})
+    end
   end
 
-  def headers_with_auth(payload)
-    token_auth_header.merge({ 'RAW_POST_DATA' => payload })
-  end
+  describe '#policy_role_counts' do
+    it 'returns the correct role counts' do
+      allow(Role).to receive(:group_and_count).and_return([
+        { kind: 'role1', count: 5 },
+        { kind: 'role2', count: 15 }
+      ])
 
-  before do
-    Slosilo["authn:rspec"] ||= Slosilo::Key.new
-    post(policies_url, env: headers_with_auth('[!group test]'))
-  end
+      counts = Monitoring::QueryHelper.instance.policy_role_counts
 
-  it 'returns policy resource counts' do
-    resource_counts = queryhelper.policy_resource_counts
-    expect(resource_counts['group']).to equal(1)
-  end
+      expect(counts).to eq({ 'role1' => 5, 'role2' => 15 })
+    end
 
-  it 'returns policy role counts' do
-    role_counts = queryhelper.policy_role_counts
-    expect(role_counts['group']).to equal(1)
-  end
+    it 'returns an empty hash if there are no roles' do
+      allow(Role).to receive(:group_and_count).and_return([])
 
+      counts = Monitoring::QueryHelper.instance.policy_role_counts
+
+      expect(counts).to eq({})
+    end
+  end
 end
